# core: remove raw SSE2/NEON implementation from matmul.cpp