--- a/modules/core/perf/opencl/perf_arithm.cpp
+++ b/modules/core/perf/opencl/perf_arithm.cpp
@@ -1076,8 +1076,14 @@ OCL_PERF_TEST_P(TransformFixture, Transform,
 
     checkDeviceMaxMemoryAllocSize(srcSize, type);
 
-    UMat src(srcSize, type), mtx(Size(4, 3), CV_32FC3), dst(srcSize, type);
-    declare.in(src, mtx, WARMUP_RNG).out(dst);
+    const float transform[] = { 0.5f,           0.f, 0.86602540378f, 128,
+                                0.f,            1.f, 0.f,            -64,
+                                0.86602540378f, 0.f, 0.5f,            32,};
+    Mat mtx(Size(4, 3), CV_32FC1, (void*)transform);
+
+    UMat src(srcSize, type), dst(srcSize, type);
+    randu(src, 0, 30);
+    declare.in(src).out(dst);
 
     OCL_TEST_CYCLE() cv::transform(src, dst, mtx);