**Merge with extra**: opencv/opencv_extra#400

### This pullrequest changes
  * use universal intrinsic instead of raw intrinsic
  * add test ```Core_TransformLargeTest``` for 8u version of transform
    * 8u SIMD version requires at least length of 16 elements  (bytes) but it seems the maximum was 4 bytes
  * may be adding test in too much, waiting for advice

<!--
force_builders=ARMv7,ARMv8,Custom
docker_image:Custom=powerpc64le
-->