@vpisarev 2017-10-02
@tomoaki0705, thank you very much! Did you compare performance of the original and the new implementation? As I see, there are some minor differences, it's not exactly 1:1 rewrite.