@tomoaki0705 2017-10-02
Thank you for your comment.Indeed, I forgot to measure the performance.
Let me report once I recover the access to the build machine.
I'm out if office this whole week, so I'll probably reply next week.