@tomoaki0705 2017-10-11
I'm now back with the measurement result, but the number doesn't look that well.

I used the test ```MatType_Length_dot``` from ```opencv_perf_core```
The summarized result is shown below

VALUE | before | after
-----|---|---
(8UC1, 32)|0.00|0.00
(8UC1, 64)|0.00|0.00
(8UC1, 128)|0.01|0.01
(8UC1, 256)|0.03|0.05
(8UC1, 512)|0.10|0.19
(8UC1, 1024)|0.50|0.88
(32SC1, 32)|0.00|0.00
(32SC1, 64)|0.01|0.01
(32SC1, 128)|0.04|0.04
(32SC1, 256)|0.17|0.17
(32SC1, 512)|0.76|0.76
(32SC1, 1024)|3.29|3.29
(32FC1, 32)|0.00|0.00
(32FC1, 64)|0.00|0.00
(32FC1, 128)|0.01|0.01
(32FC1, 256)|0.05|0.05
(32FC1, 512)|0.30|0.29
(32FC1, 1024)|2.43|2.43

So, basically as @vpisarev concerned, ```8U``` version seems getting slow, around 60-90%.
I'll try a bit more twisting the intrinsic, hoping that there are some point that every implementation can be moved to universal intrinsic