@vpisarev 2017-10-11
@tomoaki0705, thank you! So, let's pause the patch integration until 8u performance is fixed.