@seiko2plus 2017-10-12
@tomoaki0705 @vpisarev I think we should implement the following intrinsics

Left and right shifts 128bits
v_shl128(const v_reg &a, const int s)
v_shr128(const v_reg &a, const int s)

Multiply splat b and add
v_muladd_splat(const v_float32x4 &a, const v_float32x4 &b, const int spi, const v_float32x4 &c)
Returns (a * b[spi] + c)