@tomoaki0705 2017-10-24
Here's performance test from some ARM devices
Ahhhh, it works well on 32bit but some degradation on ```Aarch64```....

#### ODROID-X2
Name of Test                                           | before     | after      |x-factor 
-------------------------------------------------------------------------------------------------|------------|------------|-------- 
Mat_Transform::Size_MatType::(127x61, 8UC3)                                                      | 0.559 ms   | 0.189 ms   |  2.96   
Mat_Transform::Size_MatType::(127x61, 8SC3)                                                      | 0.612 ms   | 0.619 ms   |  0.99   
Mat_Transform::Size_MatType::(127x61, 16UC3)                                                     | 0.552 ms   | 0.317 ms   |  1.74   
Mat_Transform::Size_MatType::(127x61, 16SC3)                                                     | 0.580 ms   | 0.573 ms   |  1.01   
Mat_Transform::Size_MatType::(127x61, 32SC3)                                                     | 0.323 ms   | 0.320 ms   |  1.01   
Mat_Transform::Size_MatType::(127x61, 32FC3)                                                     | 0.236 ms   | 0.195 ms   |  1.21   
Mat_Transform::Size_MatType::(127x61, 64FC3)                                                     | 0.281 ms   | 0.281 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 8UC3)                                                     |22.818 ms   | 7.616 ms   |  3.00   
Mat_Transform::Size_MatType::(640x480, 8SC3)                                                     |24.883 ms   |25.078 ms   |  0.99   
Mat_Transform::Size_MatType::(640x480, 16UC3)                                                    |22.014 ms   |12.724 ms   |  1.73   
Mat_Transform::Size_MatType::(640x480, 16SC3)                                                    |22.445 ms   |22.343 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 32SC3)                                                    |13.002 ms   |12.992 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 32FC3)                                                    | 9.431 ms   | 8.211 ms   |  1.15   
Mat_Transform::Size_MatType::(640x480, 64FC3)                                                    |16.139 ms   |16.078 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 8UC3)                                                    |67.757 ms   |22.884 ms   |  2.96   
Mat_Transform::Size_MatType::(1280x720, 8SC3)                                                    |74.635 ms   |74.340 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 16UC3)                                                   |66.528 ms   |38.117 ms   |  1.75   
Mat_Transform::Size_MatType::(1280x720, 16SC3)                                                   |69.293 ms   |69.024 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 32SC3)                                                   |38.933 ms   |38.977 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 32FC3)                                                   |28.271 ms   |24.591 ms   |  1.15   
Mat_Transform::Size_MatType::(1280x720, 64FC3)                                                   |48.315 ms   |48.197 ms   |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 8UC3)                                                   |151.949 ms  |51.426 ms   |  2.95   
Mat_Transform::Size_MatType::(1920x1080, 8SC3)                                                   |167.751 ms  |169.523 ms  |  0.99   
Mat_Transform::Size_MatType::(1920x1080, 16UC3)                                                  |147.943 ms  |86.041 ms   |  1.72   
Mat_Transform::Size_MatType::(1920x1080, 16SC3)                                                  |151.931 ms  |153.260 ms  |  0.99   
Mat_Transform::Size_MatType::(1920x1080, 32SC3)                                                  |87.652 ms   |87.873 ms   |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 32FC3)                                                  |63.640 ms   |55.335 ms   |  1.15   
Mat_Transform::Size_MatType::(1920x1080, 64FC3)                                                  |108.213 ms  |108.506 ms  |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 8UC1)                                                   | 1.595 ms   | 1.584 ms   |  1.01   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 32FC1)                                                  | 2.351 ms   | 2.332 ms   |  1.01   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 8UC4)                                                   | 6.259 ms   | 6.260 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 32FC4)                                                  | 9.544 ms   | 9.575 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 8UC1)                                                  | 4.706 ms   | 4.708 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 32FC1)                                                 | 7.192 ms   | 7.143 ms   |  1.01   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 8UC4)                                                  |18.733 ms   |18.729 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 32FC4)                                                 |28.693 ms   |28.492 ms   |  1.01   
ScaleAdd::OCL_ScaleAddFixture::(1920x1080, 8UC1)                                                 |10.547 ms   |10.549 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1920x1080, 32FC1)                                                |16.079 ms   |16.058 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1920x1080, 8UC4)                                                 |42.113 ms   |42.105 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1920x1080, 32FC4)                                                |65.588 ms   |64.166 ms   |  1.02   
ScaleAdd::OCL_ScaleAddFixture::(3840x2160, 8UC1)                                                 |42.105 ms   |42.110 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(3840x2160, 32FC1)                                                |65.351 ms   |64.296 ms   |  1.02   
ScaleAdd::OCL_ScaleAddFixture::(3840x2160, 8UC4)                                                 |168.326 ms  |168.294 ms  |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(3840x2160, 32FC4)                                                |261.732 ms  |257.337 ms  |  1.02   
Transform::OCL_TransformFixture::(640x480, 8UC3)                                                 |21.607 ms   | 7.642 ms   |  2.83   
Transform::OCL_TransformFixture::(640x480, 8SC3)                                                 |21.586 ms   |21.582 ms   |  1.00   
Transform::OCL_TransformFixture::(640x480, 16UC3)                                                |20.550 ms   |12.738 ms   |  1.61   
Transform::OCL_TransformFixture::(640x480, 16SC3)                                                |19.832 ms   |19.826 ms   |  1.00   
Transform::OCL_TransformFixture::(640x480, 32SC3)                                                |13.036 ms   |13.015 ms   |  1.00   
Transform::OCL_TransformFixture::(640x480, 32FC3)                                                | 9.504 ms   | 8.242 ms   |  1.15   
Transform::OCL_TransformFixture::(640x480, 64FC3)                                                |16.154 ms   |16.131 ms   |  1.00   
Transform::OCL_TransformFixture::(1280x720, 8UC3)                                                |64.694 ms   |22.883 ms   |  2.83   
Transform::OCL_TransformFixture::(1280x720, 8SC3)                                                |64.713 ms   |64.719 ms   |  1.00   
Transform::OCL_TransformFixture::(1280x720, 16UC3)                                               |61.495 ms   |38.131 ms   |  1.61   
Transform::OCL_TransformFixture::(1280x720, 16SC3)                                               |59.587 ms   |59.389 ms   |  1.00   
Transform::OCL_TransformFixture::(1280x720, 32SC3)                                               |38.998 ms   |39.016 ms   |  1.00   
Transform::OCL_TransformFixture::(1280x720, 32FC3)                                               |28.372 ms   |24.625 ms   |  1.15   
Transform::OCL_TransformFixture::(1280x720, 64FC3)                                               |48.355 ms   |48.257 ms   |  1.00   
Transform::OCL_TransformFixture::(1920x1080, 8UC3)                                               |145.941 ms  |51.416 ms   |  2.84   
Transform::OCL_TransformFixture::(1920x1080, 8SC3)                                               |145.810 ms  |145.567 ms  |  1.00   
Transform::OCL_TransformFixture::(1920x1080, 16UC3)                                              |137.559 ms  |85.764 ms   |  1.60   
Transform::OCL_TransformFixture::(1920x1080, 16SC3)                                              |133.099 ms  |133.572 ms  |  1.00   
Transform::OCL_TransformFixture::(1920x1080, 32SC3)                                              |87.622 ms   |87.628 ms   |  1.00   
Transform::OCL_TransformFixture::(1920x1080, 32FC3)                                              |63.761 ms   |55.412 ms   |  1.15   
Transform::OCL_TransformFixture::(1920x1080, 64FC3)                                              |108.229 ms  |108.563 ms  |  1.00   
Transform::OCL_TransformFixture::(3840x2160, 8UC3)                                               |581.746 ms  |205.606 ms  |  2.83   
Transform::OCL_TransformFixture::(3840x2160, 8SC3)                                               |581.041 ms  |582.134 ms  |  1.00   
Transform::OCL_TransformFixture::(3840x2160, 16UC3)                                              |553.365 ms  |342.818 ms  |  1.61   
Transform::OCL_TransformFixture::(3840x2160, 16SC3)                                              |536.673 ms  |535.114 ms  |  1.00   
Transform::OCL_TransformFixture::(3840x2160, 32SC3)                                              |350.220 ms  |350.467 ms  |  1.00   
Transform::OCL_TransformFixture::(3840x2160, 32FC3)                                              |254.058 ms  |220.944 ms  |  1.15   
Transform::OCL_TransformFixture::(3840x2160, 64FC3)                                              |438.780 ms  |436.146 ms  |  1.01   
dot::MatType_Length::(8UC1, 32)                                                                  | 0.001 ms   | 0.001 ms   |  0.97   
dot::MatType_Length::(8UC1, 64)                                                                  | 0.003 ms   | 0.003 ms   |  1.01   
dot::MatType_Length::(8UC1, 128)                                                                 | 0.012 ms   | 0.012 ms   |  0.97   
dot::MatType_Length::(8UC1, 256)                                                                 | 0.058 ms   | 0.058 ms   |  1.00   
dot::MatType_Length::(8UC1, 512)                                                                 | 0.228 ms   | 0.229 ms   |  1.00   
dot::MatType_Length::(8UC1, 1024)                                                                | 1.094 ms   | 1.094 ms   |  1.00   
dot::MatType_Length::(8SC1, 32)                                                                  | 0.001 ms   | 0.001 ms   |  0.97   
dot::MatType_Length::(8SC1, 64)                                                                  | 0.003 ms   | 0.003 ms   |  1.00   
dot::MatType_Length::(8SC1, 128)                                                                 | 0.012 ms   | 0.012 ms   |  0.97   
dot::MatType_Length::(8SC1, 256)                                                                 | 0.058 ms   | 0.058 ms   |  1.00   
dot::MatType_Length::(8SC1, 512)                                                                 | 0.229 ms   | 0.229 ms   |  1.00   
dot::MatType_Length::(8SC1, 1024)                                                                | 1.091 ms   | 1.096 ms   |  1.00   
dot::MatType_Length::(16UC1, 32)                                                                 | 0.007 ms   | 0.007 ms   |  1.00   
dot::MatType_Length::(16UC1, 64)                                                                 | 0.025 ms   | 0.025 ms   |  1.00   
dot::MatType_Length::(16UC1, 128)                                                                | 0.111 ms   | 0.111 ms   |  1.00   
dot::MatType_Length::(16UC1, 256)                                                                | 0.446 ms   | 0.447 ms   |  1.00   
dot::MatType_Length::(16UC1, 512)                                                                | 1.810 ms   | 1.803 ms   |  1.00   
dot::MatType_Length::(16UC1, 1024)                                                               | 7.317 ms   | 7.314 ms   |  1.00   
dot::MatType_Length::(16SC1, 32)                                                                 | 0.007 ms   | 0.007 ms   |  1.00   
dot::MatType_Length::(16SC1, 64)                                                                 | 0.025 ms   | 0.025 ms   |  1.01   
dot::MatType_Length::(16SC1, 128)                                                                | 0.111 ms   | 0.112 ms   |  0.98   
dot::MatType_Length::(16SC1, 256)                                                                | 0.447 ms   | 0.448 ms   |  1.00   
dot::MatType_Length::(16SC1, 512)                                                                | 1.806 ms   | 1.806 ms   |  1.00   
dot::MatType_Length::(16SC1, 1024)                                                               | 7.319 ms   | 7.316 ms   |  1.00   
dot::MatType_Length::(32SC1, 32)                                                                 | 0.007 ms   | 0.007 ms   |  1.01   
dot::MatType_Length::(32SC1, 64)                                                                 | 0.027 ms   | 0.027 ms   |  0.99   
dot::MatType_Length::(32SC1, 128)                                                                | 0.122 ms   | 0.123 ms   |  1.00   
dot::MatType_Length::(32SC1, 256)                                                                | 0.487 ms   | 0.491 ms   |  0.99   
dot::MatType_Length::(32SC1, 512)                                                                | 2.045 ms   | 2.033 ms   |  1.01   
dot::MatType_Length::(32SC1, 1024)                                                               | 8.164 ms   | 8.179 ms   |  1.00   
dot::MatType_Length::(32FC1, 32)                                                                 | 0.002 ms   | 0.002 ms   |  1.00   
dot::MatType_Length::(32FC1, 64)                                                                 | 0.009 ms   | 0.009 ms   |  0.97   
dot::MatType_Length::(32FC1, 128)                                                                | 0.049 ms   | 0.049 ms   |  1.00   
dot::MatType_Length::(32FC1, 256)                                                                | 0.194 ms   | 0.195 ms   |  1.00   
dot::MatType_Length::(32FC1, 512)                                                                | 0.992 ms   | 0.994 ms   |  1.00   
dot::MatType_Length::(32FC1, 1024)                                                               | 4.381 ms   | 4.455 ms   |  0.98   

#### ODROID-C2
Name of Test                                           | before     | after      |x-factor 
-------------------------------------------------------------------------------------------------|------------|------------|-------- 
Mat_Transform::Size_MatType::(127x61, 8UC3)                                                      | 0.533 ms   | 0.180 ms   |  2.96   
Mat_Transform::Size_MatType::(127x61, 8SC3)                                                      | 0.530 ms   | 0.530 ms   |  1.00   
Mat_Transform::Size_MatType::(127x61, 16UC3)                                                     | 0.534 ms   | 0.640 ms   |  0.83   
Mat_Transform::Size_MatType::(127x61, 16SC3)                                                     | 0.529 ms   | 0.530 ms   |  1.00   
Mat_Transform::Size_MatType::(127x61, 32SC3)                                                     | 0.460 ms   | 0.459 ms   |  1.00   
Mat_Transform::Size_MatType::(127x61, 32FC3)                                                     | 0.073 ms   | 0.167 ms   |  0.44   
Mat_Transform::Size_MatType::(127x61, 64FC3)                                                     | 0.142 ms   | 0.141 ms   |  1.01   
Mat_Transform::Size_MatType::(640x480, 8UC3)                                                     |21.582 ms   | 7.421 ms   |  2.91   
Mat_Transform::Size_MatType::(640x480, 8SC3)                                                     |21.353 ms   |21.365 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 16UC3)                                                    |21.705 ms   |26.321 ms   |  0.82   
Mat_Transform::Size_MatType::(640x480, 16SC3)                                                    |21.176 ms   |21.173 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 32SC3)                                                    |18.922 ms   |18.961 ms   |  1.00   
Mat_Transform::Size_MatType::(640x480, 32FC3)                                                    | 3.168 ms   | 7.208 ms   |  0.44   
Mat_Transform::Size_MatType::(640x480, 64FC3)                                                    | 6.236 ms   | 6.209 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 8UC3)                                                    |64.302 ms   |22.170 ms   |  2.90   
Mat_Transform::Size_MatType::(1280x720, 8SC3)                                                    |63.644 ms   |63.635 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 16UC3)                                                   |65.195 ms   |78.907 ms   |  0.83   
Mat_Transform::Size_MatType::(1280x720, 16SC3)                                                   |64.765 ms   |64.765 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 32SC3)                                                   |57.143 ms   |57.124 ms   |  1.00   
Mat_Transform::Size_MatType::(1280x720, 32FC3)                                                   | 9.532 ms   |21.589 ms   |  0.44   
Mat_Transform::Size_MatType::(1280x720, 64FC3)                                                   |18.626 ms   |18.658 ms   |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 8UC3)                                                   |143.879 ms  |49.401 ms   |  2.91   
Mat_Transform::Size_MatType::(1920x1080, 8SC3)                                                   |143.637 ms  |143.645 ms  |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 16UC3)                                                  |145.647 ms  |176.398 ms  |  0.83   
Mat_Transform::Size_MatType::(1920x1080, 16SC3)                                                  |143.524 ms  |143.517 ms  |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 32SC3)                                                  |127.346 ms  |127.051 ms  |  1.00   
Mat_Transform::Size_MatType::(1920x1080, 32FC3)                                                  |21.360 ms   |48.549 ms   |  0.44   
Mat_Transform::Size_MatType::(1920x1080, 64FC3)                                                  |41.803 ms   |41.874 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 8UC1)                                                   | 0.935 ms   | 0.934 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 32FC1)                                                  | 1.611 ms   | 1.589 ms   |  1.01   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 8UC4)                                                   | 3.736 ms   | 3.743 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(640x480, 32FC4)                                                  | 6.509 ms   | 6.381 ms   |  1.02   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 8UC1)                                                  | 2.808 ms   | 2.807 ms   |  1.00   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 32FC1)                                                 | 4.856 ms   | 4.705 ms   |  1.03   
ScaleAdd::OCL_ScaleAddFixture::(1280x720, 8UC4)                                                  |11.179 ms   |11.180 ms   |  1.00   
…